Renamed GeneralPreferences to LibraryPreferences, moved reformatting and autosave to LibraryTab
