+++ src/main/java/org/jabref/cli/ArgumentProcessor.java
@@ -533,7 +533,7 @@ public class ArgumentProcessor {
             try (AtomicFileWriter fileWriter = new AtomicFileWriter(Path.of(subName), StandardCharsets.UTF_8)) {
                 BibWriter bibWriter = new BibWriter(fileWriter, OS.NEWLINE);
                 SaveConfiguration saveConfiguration = new SaveConfiguration()
-                        .withReformatOnSave(preferencesService.getExportPreferences().shouldAlwaysReformatOnSave());
+                        .withReformatOnSave(preferencesService.getGeneralPreferences().shouldAlwaysReformatOnSave());
 
                 BibDatabaseWriter databaseWriter = new BibtexDatabaseWriter(
                         bibWriter,
