+++ src/main/java/org/jabref/gui/preferences/export/ExportTab.fxml
@@ -1,9 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.CheckBox?>
 <?import javafx.scene.control.Label?>
-<?import javafx.scene.layout.HBox?>
 <?import javafx.scene.layout.VBox?>
 <?import org.jabref.gui.commonfxcontrols.SaveOrderConfigPanel?>
 <fx:root spacing="10.0" type="VBox"
@@ -13,11 +10,4 @@
 
     <Label styleClass="sectionHeader" text="%Export sort order"/>
     <SaveOrderConfigPanel fx:id="exportOrderPanel"/>
-
-    <Label styleClass="sectionHeader" text="%Saving"/>
-    <CheckBox fx:id="alwaysReformatBib" text="%Always reformat BIB file on save and export"/>
-    <HBox alignment="CENTER_LEFT" spacing="10.0">
-        <CheckBox fx:id="autosaveLocalLibraries" text="%Autosave local libraries"/>
-        <Button fx:id="autosaveLocalLibrariesHelp"/>
-    </HBox>
 </fx:root>
