+++ src/main/java/org/jabref/gui/preferences/library/LibraryTab.fxml
@@ -1,9 +1,13 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.CheckBox?>
 <?import javafx.scene.control.ComboBox?>
 <?import javafx.scene.control.Label?>
 <?import javafx.scene.layout.ColumnConstraints?>
 <?import javafx.scene.layout.GridPane?>
+<?import javafx.scene.layout.HBox?>
 <?import javafx.scene.layout.VBox?>
 <fx:root spacing="10.0" type="VBox"
          xmlns="http://javafx.com/javafx" xmlns:fx="http://javafx.com/fxml"
@@ -18,4 +22,14 @@
         <Label text="%Default library mode"/>
         <ComboBox fx:id="biblatexMode" prefWidth="200.0" GridPane.columnIndex="1"/>
     </GridPane>
+
+    <Label styleClass="sectionHeader" text="%Saving"/>
+    <CheckBox fx:id="alwaysReformatBib" text="%Always reformat BIB file on save and export"/>
+    <HBox alignment="CENTER_LEFT" spacing="4.0">
+        <CheckBox fx:id="autosaveLocalLibraries" text="%Autosave local libraries"/>
+        <Button fx:id="autosaveLocalLibrariesHelp"/>
+        <padding>
+            <Insets top="-8.0"/>
+        </padding>
+    </HBox>
 </fx:root>
