+++ src/main/java/org/jabref/gui/preferences/library/LibraryTabViewModel.java
@@ -1,7 +1,9 @@
 package org.jabref.gui.preferences.library;
 
+import javafx.beans.property.BooleanProperty;
 import javafx.beans.property.ListProperty;
 import javafx.beans.property.ObjectProperty;
+import javafx.beans.property.SimpleBooleanProperty;
 import javafx.beans.property.SimpleListProperty;
 import javafx.beans.property.SimpleObjectProperty;
 import javafx.collections.FXCollections;
@@ -9,7 +11,7 @@ import javafx.collections.FXCollections;
 import org.jabref.gui.DialogService;
 import org.jabref.gui.preferences.PreferenceTabViewModel;
 import org.jabref.model.database.BibDatabaseMode;
-import org.jabref.preferences.GeneralPreferences;
+import org.jabref.preferences.LibraryPreferences;
 import org.jabref.preferences.TelemetryPreferences;
 
 public class LibraryTabViewModel implements PreferenceTabViewModel {
@@ -17,21 +19,30 @@ public class LibraryTabViewModel implements PreferenceTabViewModel {
     private final ListProperty<BibDatabaseMode> bibliographyModeListProperty = new SimpleListProperty<>();
     private final ObjectProperty<BibDatabaseMode> selectedBiblatexModeProperty = new SimpleObjectProperty<>();
 
-    private final DialogService dialogService;
-    private final GeneralPreferences generalPreferences;
+    private final BooleanProperty alwaysReformatBibProperty = new SimpleBooleanProperty();
+    private final BooleanProperty autosaveLocalLibraries = new SimpleBooleanProperty();
 
-    public LibraryTabViewModel(DialogService dialogService, GeneralPreferences generalPreferences, TelemetryPreferences telemetryPreferences) {
+    private final DialogService dialogService;
+    private final LibraryPreferences libraryPreferences;
+
+    public LibraryTabViewModel(DialogService dialogService, LibraryPreferences libraryPreferences, TelemetryPreferences telemetryPreferences) {
         this.dialogService = dialogService;
-        this.generalPreferences = generalPreferences;
+        this.libraryPreferences = libraryPreferences;
     }
 
     public void setValues() {
         bibliographyModeListProperty.setValue(FXCollections.observableArrayList(BibDatabaseMode.values()));
-        selectedBiblatexModeProperty.setValue(generalPreferences.getDefaultBibDatabaseMode());
+        selectedBiblatexModeProperty.setValue(libraryPreferences.getDefaultBibDatabaseMode());
+
+        alwaysReformatBibProperty.setValue(libraryPreferences.shouldAlwaysReformatOnSave());
+        autosaveLocalLibraries.setValue(libraryPreferences.shouldAutoSave());
     }
 
     public void storeSettings() {
-        generalPreferences.setDefaultBibDatabaseMode(selectedBiblatexModeProperty.getValue());
+        libraryPreferences.setDefaultBibDatabaseMode(selectedBiblatexModeProperty.getValue());
+
+        libraryPreferences.setAlwaysReformatOnSave(alwaysReformatBibProperty.getValue());
+        libraryPreferences.setAutoSave(autosaveLocalLibraries.getValue());
     }
 
     public ListProperty<BibDatabaseMode> biblatexModeListProperty() {
@@ -40,5 +51,13 @@ public class LibraryTabViewModel implements PreferenceTabViewModel {
 
     public ObjectProperty<BibDatabaseMode> selectedBiblatexModeProperty() {
         return this.selectedBiblatexModeProperty;
+    }
+
+    public BooleanProperty alwaysReformatBibProperty() {
+        return alwaysReformatBibProperty;
+    }
+
+    public BooleanProperty autosaveLocalLibrariesProperty() {
+        return autosaveLocalLibraries;
     }
 }
