+++ src/main/java/org/jabref/logic/autosaveandbackup/BackupManager.java
@@ -227,7 +227,7 @@ public class BackupManager {
         // code similar to org.jabref.gui.exporter.SaveDatabaseAction.saveDatabase
         SaveConfiguration saveConfiguration = new SaveConfiguration()
                 .withMakeBackup(false)
-                .withReformatOnSave(preferences.getExportPreferences().shouldAlwaysReformatOnSave());
+                .withReformatOnSave(preferences.getGeneralPreferences().shouldAlwaysReformatOnSave());
 
         Charset encoding = bibDatabaseContext.getMetaData().getEncoding().orElse(StandardCharsets.UTF_8);
         // We want to have successful backups only
