+++ src/main/java/org/jabref/preferences/LibraryPreferences.java
@@ -1,0 +1,59 @@
+package org.jabref.preferences;
+
+import javafx.beans.property.BooleanProperty;
+import javafx.beans.property.ObjectProperty;
+import javafx.beans.property.SimpleBooleanProperty;
+import javafx.beans.property.SimpleObjectProperty;
+
+import org.jabref.model.database.BibDatabaseMode;
+
+public class LibraryPreferences {
+
+    private final ObjectProperty<BibDatabaseMode> defaultBibDatabaseMode;
+    private final BooleanProperty alwaysReformatOnSave;
+    private final BooleanProperty autoSave;
+
+    public LibraryPreferences(BibDatabaseMode defaultBibDatabaseMode,
+                              boolean alwaysReformatOnSave,
+                              boolean autoSave) {
+        this.defaultBibDatabaseMode = new SimpleObjectProperty<>(defaultBibDatabaseMode);
+        this.alwaysReformatOnSave = new SimpleBooleanProperty(alwaysReformatOnSave);
+        this.autoSave = new SimpleBooleanProperty(autoSave);
+    }
+
+    public BibDatabaseMode getDefaultBibDatabaseMode() {
+        return defaultBibDatabaseMode.get();
+    }
+
+    public ObjectProperty<BibDatabaseMode> defaultBibDatabaseModeProperty() {
+        return defaultBibDatabaseMode;
+    }
+
+    public void setDefaultBibDatabaseMode(BibDatabaseMode defaultBibDatabaseMode) {
+        this.defaultBibDatabaseMode.set(defaultBibDatabaseMode);
+    }
+
+    public boolean shouldAlwaysReformatOnSave() {
+        return alwaysReformatOnSave.get();
+    }
+
+    public BooleanProperty alwaysReformatOnSaveProperty() {
+        return alwaysReformatOnSave;
+    }
+
+    public void setAlwaysReformatOnSave(boolean alwaysReformatOnSave) {
+        this.alwaysReformatOnSave.set(alwaysReformatOnSave);
+    }
+
+    public boolean shouldAutoSave() {
+        return autoSave.get();
+    }
+
+    public BooleanProperty autoSaveProperty() {
+        return autoSave;
+    }
+
+    public void setAutoSave(boolean shouldAutoSave) {
+        this.autoSave.set(shouldAutoSave);
+    }
+}
