+++ src/test/java/org/jabref/logic/crawler/StudyRepositoryTest.java
@@ -27,7 +27,7 @@ import org.jabref.model.entry.types.StandardEntryType;
 import org.jabref.model.study.FetchResult;
 import org.jabref.model.study.QueryResult;
 import org.jabref.model.util.DummyFileUpdateMonitor;
-import org.jabref.preferences.GeneralPreferences;
+import org.jabref.preferences.LibraryPreferences;
 import org.jabref.preferences.PreferencesService;
 
 import org.junit.jupiter.api.BeforeEach;
@@ -47,7 +47,7 @@ class StudyRepositoryTest {
     private static final String NON_EXISTING_DIRECTORY = "nonExistingTestRepositoryDirectory";
     CitationKeyPatternPreferences citationKeyPatternPreferences;
     PreferencesService preferencesService;
-    GeneralPreferences generalPreferences;
+    LibraryPreferences libraryPreferences;
     ImportFormatPreferences importFormatPreferences;
     SaveConfiguration saveConfiguration;
     BibEntryTypesManager entryTypesManager;
@@ -64,7 +64,7 @@ class StudyRepositoryTest {
      */
     @BeforeEach
     public void setUpMocks() throws Exception {
-        generalPreferences = mock(GeneralPreferences.class, Answers.RETURNS_DEEP_STUBS);
+        libraryPreferences = mock(LibraryPreferences.class, Answers.RETURNS_DEEP_STUBS);
         saveConfiguration = mock(SaveConfiguration.class, Answers.RETURNS_DEEP_STUBS);
         importFormatPreferences = mock(ImportFormatPreferences.class, Answers.RETURNS_DEEP_STUBS);
         preferencesService = mock(PreferencesService.class, Answers.RETURNS_DEEP_STUBS);
